agent-service/requirements.txt
# Fast, modern web framework for building APIs
fastapi

# Lightning-fast ASGI server, to run our FastAPI application
uvicorn

# The core MCP library we will use for our agentic tools
fastmcp

# A fast Python package installer, recommended by the FastMCP docs
uv

# Google Cloud SDK for Vertex AI
google-cloud-aiplatform

# LangChain integration for Vertex AI, used by CrewAI
langchain-google-vertexai

# Utility to load environment variables from a .env file for local development
python-dotenv

# For making HTTP requests to fetch articles
requests

# For building multi-agent AI systems
crewai
crewai[tools]

# PostgreSQL database driver for Python
psycopg2-binary

# libraries for robust HTML parsing
beautifulsoup4
readability-lxml

# For RabbitMQ communication
pika
supervisor